added getall method. Feat - to avoid creating separate table for tags, made the tags a string in entity while keeping list in Dto. added a method to convert tags separated by commas into list item. todo -need to create a methode to convert tags back into strings when adding items

+++ src/main/java/edu/icet/ecom/repository/BlogRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlogRepository extends JpaRepository<BlogEntity,Long> {
+public interface BlogRepository extends JpaRepository<BlogEntity,Integer> {
 }
